refactor: nns prefix
