+++ frontend/src/lib/components/neuron-detail/actions/SplitNnsNeuronButton.svelte
