+++ frontend/src/tests/lib/components/neuron-detail/actions/SplitNnsNeuronButton.spec.ts
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 
-import SplitNeuronButton from "$lib/components/neuron-detail/actions/SplitNeuronButton.svelte";
+import SplitNeuronButton from "$lib/components/neuron-detail/actions/SplitNnsNeuronButton.svelte";
 import { fireEvent, render } from "@testing-library/svelte";
 import en from "../../../../mocks/i18n.mock";
 import { mockFullNeuron, mockNeuron } from "../../../../mocks/neurons.mock";
